move example to own component
