+++ src/components/HamletComponents/index.js
@@ -1,12 +1,10 @@
 import React, { useState, useEffect } from "react";
 import Admonition from "react-admonitions";
-import Tabs from "@theme/Tabs";
-import TabItem from "@theme/TabItem";
-import Highlight, { defaultProps } from "prism-react-renderer";
 
 import axios from "axios";
 
 import "./styles.css";
+import HamletExample from "@site/src/components/HamletExample";
 
 const schemaUrl =
   "https://hamlet-docs-git-add-schema-documents.hamlet.now.sh/schema/latest/blueprint/component.json";
@@ -16,16 +14,6 @@ const filterAttributerObjects = [
   "Instances",
   "additionalProperties",
 ];
-
-const examplejson = {
-  Tiers: {
-    tiername: {
-      Components: {
-        componentname: {},
-      },
-    },
-  },
-};
 
 const getAttributeStructure = (attributes) => {
   let results = [];
@@ -290,28 +278,7 @@ function HamletComponent(props) {
             {props.name + " Component"}
           </div>
           <div className="col col--8">
-            <Tabs
-              defaultValue="json"
-              values={[
-                { label: "JSON", value: "json" },
-                { label: "YAML", value: "yaml" },
-              ]}
-            >
-              <TabItem value="json">
-                <Highlight {...defaultProps} code="" language="json">
-                  {(highlight) => (
-                    <pre>{JSON.stringify(examplejson, null, 2)}</pre>
-                  )}
-                </Highlight>
-              </TabItem>
-              <TabItem value="yaml">
-                <Highlight {...defaultProps} code="" language="yaml">
-                  {(highlight) => (
-                    <pre>{JSON.stringify({ test: "test2" }, null, 2)}</pre>
-                  )}
-                </Highlight>
-              </TabItem>
-            </Tabs>
+            <HamletExample />
           </div>
         </div>
       </div>
